refactor: pkg/args/internal/text → internal/text
