+++ internal/text/ast.go
