+++ internal/text/text.go
